Made plugin more robust against different project types.
When the plugin is executed in a project of type pom it throws an
exception because it cannot find the source directory. This is now
suppressed, and no data is generated for this project.

The use case is that there can be multi-module projects defining
multiple tag libs. Then it makes sense to have a parent project (of
type pom) which configures the plugin. The plugin is then applied to
all taglib sub projects. With this patch, this use case is now
supported.